+++ CHANGELOG.md
@@ -38,6 +38,9 @@
   This requires main menu commands to implement the `mainmenu_commands_v3`
   service interface in the foobar2000 SDK.
 
+- A ‘Cut panel’ command was added to the Layout tree context menu in
+  Preferences. [#1302](https://github.com/reupen/columns_ui/pull/1302)
+
 ### Bug fixes
 
 - The built-in spectrum analyser visualisation now uses the same frequency scale